[refactor] moving Label and associated method into asm_utility module
