+++ metalibm_core/code_generation/asm_utility.py
@@ -1,0 +1,53 @@
+# -*- coding: utf-8 -*-
+
+###############################################################################
+# This file is part of metalibm (https://github.com/kalray/metalibm)
+###############################################################################
+# MIT License
+#
+# Copyright (c) 2020 Kalray
+#
+# Permission is hereby granted, free of charge, to any person obtaining a copy
+# of this software and associated documentation files (the "Software"), to deal
+# in the Software without restriction, including without limitation the rights
+# to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+# copies of the Software, and to permit persons to whom the Software is
+# furnished to do so, subject to the following conditions:
+#
+# The above copyright notice and this permission notice shall be included in all
+# copies or substantial portions of the Software.
+#
+# THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+# IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+# FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+# AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+# LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+# OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+# SOFTWARE.
+###############################################################################
+# created:          May  4rd, 2020
+# last-modified:    May  4rd, 2020
+#
+# author(s): Nicolas Brunie (nicolas.brunie@kalray.eu)
+###############################################################################
+
+from metalibm_core.code_generation.code_object import MultiSymbolTable
+
+
+class Label(object):
+    """ Label (tag for code position) object """
+    def __init__(self, name):
+        self.name = name
+
+def get_free_label_name(code_object, prefix):
+    """ generate a new label name (previously unused) """
+    return code_object.get_free_symbol_name(
+        MultiSymbolTable.LabelSymbol, None, Label,
+        prefix=prefix,
+        declare=True)
+
+def append_label(code_object, label):
+    """ append a new label location at the end of code_object """
+    code_object.close_level(footer="", cr="")
+    code_object << label << ":"
+    code_object.open_level(header="") #, extra_shared_tables=[MultiSymbolTable.VariableSymbol])
